services/sync-worker/requirements.txt
redis==5.2.1
httpx==0.28.1
python-dotenv==1.0.1
